fix button list order
