+++ src/components/table-collapse/index.js
@@ -1,28 +1,27 @@
-import React, { useState, useEffect } from "react";
-import { makeStyles } from "@material-ui/core/styles";
+import React, { useState } from "react";
 import Table from "@material-ui/core/Table";
 import TableBody from "@material-ui/core/TableBody";
 import TableCell from "@material-ui/core/TableCell";
-//import TableContainer from "@material-ui/core/TableContainer";
-//import TableHead from "@material-ui/core/TableHead";
 import TableRow from "@material-ui/core/TableRow";
-//import Paper from "@material-ui/core/Paper";
 import Button from '@material-ui/core/Button';
 import Collapse from '@material-ui/core/Collapse';
 import IconButton from '@material-ui/core/IconButton';
 import KeyboardArrowDownIcon from '@material-ui/icons/KeyboardArrowDown';
 import KeyboardArrowUpIcon from '@material-ui/icons/KeyboardArrowUp';
+import { updateOrderStatus } from "../products/updateOrder.js"
 
-
-/*const useStyles = makeStyles({
-  table: {
-    minWidth: 500
-  }
-});*/
 
 export const TableCollapse = (props) => {
-  //const classes = useStyles();
   const [open, setOpen] = useState(false);
+
+  const [status, setStatus] = useState(props.orderDone.status);
+  
+  const orderStatus = {
+    "pending": "Pendente",
+    "doing": "Em preparo",
+    "done": "Pedido pronto",
+    "delivery": "Entregue"
+  }
 
   return (
     <React.Fragment>
@@ -34,8 +33,8 @@ export const TableCollapse = (props) => {
             </IconButton> Itens do pedido
           </TableCell>
           <TableCell>Mesa:{props.orderDone.table}</TableCell>
-          <TableCell colSpan={3} align="right">
-            <Button variant="contained" color="primary">Confirmar Entrega</Button>
+          <TableCell align="right">
+            <Button variant="contained" color="primary" onClick={() => updateOrderStatus("delivery", props.orderDone.id, setStatus)}>Confirmar Entrega</Button>
           </TableCell>
         </TableRow>
       </TableBody>
